# Improve performance around serialization