--- a/include/Basic/ASerializable.hpp
+++ b/include/Basic/ASerializable.hpp
@@ -120,7 +120,7 @@ bool ASerializable::_recordWrite(std::ostream& os,
       if (title.empty())
         os << STRING_NA << " ";
       else
-        os << STRING_NA << " # " << title << std::endl;
+        os << STRING_NA << " # " << title << '\n';
     }
     else
     {
@@ -129,7 +129,7 @@ bool ASerializable::_recordWrite(std::ostream& os,
       if (title.empty())
         os << val << " ";
       else
-        os << val << " # " << title << std::endl;
+        os << val << " # " << title << '\n';
       os.precision(prec);
     }
   }
@@ -144,7 +144,7 @@ bool ASerializable::_recordWriteVec(std::ostream& os,
   if (os.good())
   {
     if (!title.empty())
-      os << "# " << title << std::endl;
+      os << "# " << title << '\n';
 
     int prec = os.precision();
     os.precision(15);
@@ -155,7 +155,7 @@ bool ASerializable::_recordWriteVec(std::ostream& os,
       else
         os << val << " ";
     }
-    os << std::endl;
+    os << '\n';
     os.precision(prec);
   }
   return os.good();